requirements.txt
PySide6
python-dotenv
python-dateutil
matplotlib
openpyxl
pywin32
google-generativeai
